# InvServer.Api/wwwroot/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InvServer API",
    "description": "Comprehensive API for Inventory Management, Workflow Governance, and Idempotent Fulfillment.\n\nKey Features:\n- **Idempotency**: All fulfillment operations require `X-Idempotency-Key`.\n- **Workflow**: Governed lifecycle (Draft -> In Workflow -> Fulfillment).\n- **RBAC**: Scope-based access control (OWN, DEPT, WAREHOUSE, GLOBAL).\n\n(Updated to reflect latest Admin/Workflow/Inventory controller changes.)",
    "version": "1.2.0",
    "contact": {
      "name": "API Support",
      "email": "support@invserver.local"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object"
          },
          "correlationId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "ApiResponse_Generic": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "correlationId": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
      "ApiResponse_AuthResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthResponse"
          },
          "correlationId": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "jdoe"
          },
          "password": {
            "type": "string",
            "example": "SecurePass123!"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string",
            "description": "Empty if using HttpOnly cookie flow"
          },
          "user": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "integer",
                "format": "int64"
              },
              "username": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "departments": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
...
  },
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Returns an access token and sets a standard `refreshToken` HttpOnly cookie.",
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh Token",
        "description": "Uses the `refreshToken` cookie to issue a new access token.",
        "operationId": "Refresh",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid/Expired Token"
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Current User",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        }
      }
    },
...
  }
}